refactor container and cleanup drag layers
