+++ frontend/src/Editor/Container.jsx
@@ -1,14 +1,12 @@
 /* eslint-disable import/no-named-as-default */
 import React, { useCallback, useState, useEffect, useRef } from 'react';
 import cx from 'classnames';
-import { v4 as uuidv4 } from 'uuid';
 import { useDrop, useDragLayer } from 'react-dnd';
 import { ItemTypes } from './ItemTypes';
 import { DraggableBox } from './DraggableBox';
-import { snapToGrid as doSnapToGrid } from './snapToGrid';
 import update from 'immutability-helper';
 import { componentTypes } from './WidgetManager/components';
-import { computeComponentName, resolveReferences } from '@/_helpers/utils';
+import { resolveReferences } from '@/_helpers/utils';
 import useRouter from '@/_hooks/use-router';
 import Comments from './Comments';
 import { commentsService } from '@/_services';
@@ -16,7 +14,7 @@ import config from 'config';
 import Spinner from '@/_ui/Spinner';
 import { useHotkeys } from 'react-hotkeys-hook';
 import produce from 'immer';
-import { addComponents } from '@/_helpers/appUtils';
+import { addComponents, addNewWidgetToTheEditor } from '@/_helpers/appUtils';
 
 export const Container = ({
   canvasWidth,
@@ -184,65 +182,31 @@ export const Container = ({
           return undefined;
         }
 
-        // let layouts = item['layouts'];
-        // const currentLayoutOptions = layouts ? layouts[item.currentLayout] : {};
-
-        let componentData = {};
-        let componentMeta = {};
-        let id = item.id;
-
-        let left = 0;
-        let top = 0;
-
         const canvasBoundingRect = document.getElementsByClassName('real-canvas')[0].getBoundingClientRect();
-
-        //  This is a new component
-        componentMeta = componentTypes.find((component) => component.component === item.component.component);
+        const componentMeta = componentTypes.find((component) => component.component === item.component.component);
         console.log('adding new component');
-        componentData = JSON.parse(JSON.stringify(componentMeta));
-        componentData.name = computeComponentName(componentData.component, boxes);
 
-        const offsetFromTopOfWindow = canvasBoundingRect.top;
-        const offsetFromLeftOfWindow = canvasBoundingRect.left;
-        const currentOffset = monitor.getSourceClientOffset();
-        const initialClientOffset = monitor.getInitialClientOffset();
-        const delta = monitor?.getDifferenceFromInitialOffset();
-        left = Math.round(currentOffset?.x + currentOffset?.x * (1 - zoomLevel) - offsetFromLeftOfWindow);
-        top = Math.round(
-          initialClientOffset?.y - 10 + delta.y + initialClientOffset?.y * (1 - zoomLevel) - offsetFromTopOfWindow
+        const newComponent = addNewWidgetToTheEditor(
+          componentMeta,
+          monitor,
+          boxes,
+          canvasBoundingRect,
+          item.currentLayout,
+          {},
+          snapToGrid,
+          zoomLevel
         );
-
-        id = uuidv4();
-        const bundingRect = document.getElementsByClassName('canvas-area')[0].getBoundingClientRect();
-        const canvasWidth = bundingRect?.width;
-
-        if (snapToGrid) {
-          [left, top] = doSnapToGrid(canvasWidth, left, top);
-        }
-
-        left = (left * 100) / canvasWidth;
-
-        if (item.currentLayout === 'mobile') {
-          componentData.definition.others.showOnDesktop.value = false;
-          componentData.definition.others.showOnMobile.value = true;
-        }
-
-        const width = componentMeta.defaultSize.width;
 
         setBoxes({
           ...boxes,
-          [id]: {
-            component: componentData,
+          [newComponent.id]: {
+            component: newComponent.component,
             layouts: {
-              [item.currentLayout]: {
-                top,
-                left,
-                width,
-                height: componentMeta.defaultSize.height,
-              },
+              ...newComponent.layout,
             },
           },
         });
+
         return undefined;
       },
     }),
