--- frontend/src/Editor/snapToGrid.js
@@ -1,7 +1,0 @@
-export function snapToGrid(canvasWidth, x, y) {
-  const gridX = canvasWidth / 43;
-
-  const snappedX = Math.round(x / gridX) * gridX;
-  const snappedY = Math.round(y / 10) * 10;
-  return [snappedX, snappedY];
-}
